+++ src/lib.rs
@@ -95,7 +95,8 @@ pub use crate::media::{
 pub use crate::pagination::PageRange;
 pub use crate::server::{PlexServer, ServerIdentity};
 pub use crate::traits::{
-    EditField, EditSummary, EditTitle, FieldValue, PlayedUnplayed, PlexObject, Ratable,
+    EditField, EditSummary, EditTags, EditTitle, FieldValue, HasCollections, HasGenres,
+    PlayedUnplayed, PlexObject, Ratable,
 };
 pub use crate::uri::PlexUri;
 pub use crate::util::{